backend/api-gateway/requirements.txt
fastapi
uvicorn
pydantic
python-jose
python-dotenv
httpx
pytest
